Renamed widget to avoid it's detection by adblockers
