+++ EventAnalytics/.gitignore
@@ -3,3 +3,4 @@ node_modules/
 tests/testProject/
 *.log
 .env
+.idea
