+++ EventAnalytics/NlyticsEvent.tsx
@@ -1,6 +1,6 @@
-import { nanoid } from "nanoid";
+import { v4 as uuidv4 } from "uuid";
 import { createElement, Fragment, useEffect, useRef } from "react";
-import { AnalyticsEventContainerProps, ClassNameToTrackType } from "../typings/AnalyticsEventProps";
+import { NlyticsEventContainerProps, ClassNameToTrackType } from "./NlyticsEventProps";
 import PubSub from "pubsub-js";
 
 const eventListenerName = "CLASSNAME_MENDIX_LISTENER";
@@ -9,8 +9,8 @@ const stringifyMe = (payLoad: any) => {
     return JSON.stringify(payLoad);
 };
 
-export const AnalyticsEvent = (props: AnalyticsEventContainerProps) => {
-    const _id = useRef(nanoid());
+export const NlyticsEvent = (props: NlyticsEventContainerProps) => {
+    const _id = useRef(uuidv4());
     const dispatchClickEvent = (className: ClassNameToTrackType, id: string) => {
         const payLoad = {
             ...className,
@@ -29,7 +29,7 @@ export const AnalyticsEvent = (props: AnalyticsEventContainerProps) => {
                     const foundClassName = Array.from(q.classList).find(classItem =>
                         classItem.includes(className.className)
                     );
-                    const _id = nanoid();
+                    const _id = uuidv4();
                     const uu: ClassNameToTrackType = {
                         ...className,
                         friendlyName: foundClassName as string,
@@ -40,7 +40,7 @@ export const AnalyticsEvent = (props: AnalyticsEventContainerProps) => {
             } else {
                 const classFound = document.getElementsByClassName(className.className)[0];
                 if (classFound) {
-                    const _id = nanoid();
+                    const _id = uuidv4();
                     classFound.addEventListener("click", () => dispatchClickEvent(className, _id)) as any;
                     return classFound;
                 }
