+++ EventAnalytics/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <package xmlns="http://www.mendix.com/package/1.0/">
-    <clientModule name="AnalyticsEvent" version="1.0.0" xmlns="http://www.mendix.com/clientModule/1.0/">
+    <clientModule name="NlyticsEvent" version="1.0.0" xmlns="http://www.mendix.com/clientModule/1.0/">
         <widgetFiles>
-            <widgetFile path="AnalyticsEvent.xml" />
+            <widgetFile path="NlyticsEvent.xml" />
         </widgetFiles>
         <files>
             <file path="mendix/analyticsevent" />
